OPS HLS Library side restructure, seperating common components into seperate directory.
