+++ ops/hls/include/device/ops_hls_kernel_support.h
@@ -9,10 +9,12 @@
   */
 
 #include <ap_int.h>
+#include "../../common/include/ops_hls_defs.hpp"
+#include "../../common/include/ops_hls_utils.hpp"
+#include "../../common/include/ops_hls_common_memconfig.hpp"
 #include "../../L1/include/ops_hls_datamover.hpp"
-#include "../../L1/include/ops_hls_defs.hpp"
 #include "../../L1/include/ops_hls_stencil_core_v2.hpp"
-#include "../../L1/include/ops_hls_utils.hpp"
+
 // typedef struct ops_hls_core
 // {
 //     ops::hls::FPGA* fpga_handle;
